V4 merged express

+++ ai_exportpdf/utils.py
@@ -89,6 +89,7 @@ def convertiopdf2docx(id ,language,ocr = None ):
     response_status = requests.post(url='https://api.convertio.co/convert' , data=json.dumps(data)).json()
     if response_status['status'] == 'error':
         txt_field_obj.status = "ERROR"
+        txt_field_obj.pdf_api_use = "FileCorrupted"
         txt_field_obj.save()
         ###retain cred if error
         file_format,page_length =  file_pdf_check(fp,id)
@@ -124,6 +125,7 @@ def convertiopdf2docx(id ,language,ocr = None ):
             # end = time.time()
             else:
                 txt_field_obj.status = "ERROR"
+                txt_field_obj.pdf_api_use = "FileCorrupted"
                 txt_field_obj.save()
                 file_format,page_length =  file_pdf_check(fp,id)
                 # file_format,page_length = pdf_text_check(fp)
@@ -177,6 +179,7 @@ def ai_export_pdf(id): # , file_language , file_name , file_path
     except:
         end = time.time()
         txt_field_obj.status = "ERROR"
+        txt_field_obj.pdf_api_use = "FileCorrupted"
         txt_field_obj.save()
         ###retain cred if error
         file_format,page_length =  file_pdf_check(fp , id ) 
@@ -208,7 +211,7 @@ def para_creation_from_ocr(texts):
 
 import PyPDF2
 from rest_framework import serializers
-def file_pdf_check(file_path,id): 
+def file_pdf_check(file_path,pdf_id): 
     try:
         pdfdoc = PyPDF2.PdfReader(file_path)
         pdf_check = {0:'ocr',1:'text'}
@@ -224,10 +227,16 @@ def file_pdf_check(file_path,id):
                 pdf_check_list.append(0)
         return [pdf_check.get(max(pdf_check_list)) , len(pdfdoc.pages)]
     except:
-        file_details = Ai_PdfUpload.objects.get(id = id)
-        file_details.delete()
-        raise serializers.ValidationError({'msg':'pdf_corrupted'}, 
-                                          code =400)
+        if pdf_id:
+            file_details = Ai_PdfUpload.objects.get(id = pdf_id)
+            file_details.delete()
+            # return None,None
+            # file_details.status = "FileCorrupted"
+            # file_details.save()
+            raise serializers.ValidationError({'msg':'pdf_corrupted'}, 
+                                            code =400)
+        else:
+            return None,None
